requirements.txt
fastapi 
uvicorn
anthropic
requests
pillow
